+++ lib/glimmer/libui/control_proxy/form_proxy.rb
@@ -39,6 +39,7 @@ module Glimmer
         end
         
         def destroy_child(child)
+          child.deregister_all_custom_listeners
           ::LibUI.send("form_delete", @libui, children.index(child))
           ControlProxy.control_proxies.delete(child)
         end