+++ lib/glimmer/libui/control_proxy/menu_item_proxy/quit_menu_item_proxy.rb
@@ -56,6 +56,11 @@ module Glimmer
               end
             end
           end
+          
+          def destroy
+            @on_clicked_listeners&.clear
+            super
+          end
         
           private